requirements.txt
blinker==1.9.0
click==8.3.2
Flask==3.1.3
Flask-Scss==0.5
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.2.0
pyScss==1.4.0
six==1.17.0
Werkzeug==3.1.8
